chore: Fix SSR build (#14331)
This PR fixes the build of the SSR setup lib by removing a cyclic dependency.
