+++ core-libs/setup/ng-package.json
@@ -5,6 +5,7 @@
     "entryFile": "public_api.ts",
     "umdModuleIds": {
       "@spartacus/checkout": "checkout",
+      "@spartacus/checkout/root": "checkoutRoot",
       "@spartacus/core": "core",
       "@spartacus/storefront": "storefront",
       "@spartacus/user": "user"
