+++ core-libs/setup/ssr/engine-decorator/ng-express-engine-decorator.ts
@@ -1,13 +1,10 @@
-import { StaticProvider } from '@angular/core';
 import { NgSetupOptions } from '@nguniversal/express-engine';
-import { REQUEST } from '@nguniversal/express-engine/tokens';
-import { SERVER_REQUEST_ORIGIN, SERVER_REQUEST_URL } from '@spartacus/core';
-import { Request } from 'express';
 import {
   OptimizedSsrEngine,
   SsrCallbackFn,
 } from '../optimized-engine/optimized-ssr-engine';
 import { SsrOptimizationOptions } from '../optimized-engine/ssr-optimization-options';
+import { getServerRequestProviders } from '../providers/ssr-providers';
 
 export type NgExpressEngineInstance = (
   filePath: string,
@@ -60,47 +57,4 @@ export function decorateExpressEngine(
           .engineInstance
       : engineInstance;
   };
-}
-
-/**
- * Returns Spartacus providers to be passed to the Angular express engine (in SSR)
- *
- * @param options
- */
-export function getServerRequestProviders(): StaticProvider[] {
-  return [
-    {
-      provide: SERVER_REQUEST_URL,
-      useFactory: getRequestUrl,
-      deps: [REQUEST],
-    },
-    {
-      provide: SERVER_REQUEST_ORIGIN,
-      useFactory: getRequestOrigin,
-      deps: [REQUEST],
-    },
-  ];
-}
-
-export function getRequestUrl(req: Request): string {
-  return getRequestOrigin(req) + req.originalUrl;
-}
-
-export function getRequestOrigin(req: Request): string {
-  // If express is resolving and trusting X-Forwarded-Host, we want to take it
-  // into an account to properly generate request origin.
-  const trustProxyFn = req.app.get('trust proxy fn');
-  let forwardedHost = req.get('X-Forwarded-Host');
-  if (forwardedHost && trustProxyFn(req.connection.remoteAddress, 0)) {
-    if (forwardedHost.indexOf(',') !== -1) {
-      // Note: X-Forwarded-Host is normally only ever a
-      //       single value, but this is to be safe.
-      forwardedHost = forwardedHost
-        .substring(0, forwardedHost.indexOf(','))
-        .trimRight();
-    }
-    return req.protocol + '://' + forwardedHost;
-  } else {
-    return req.protocol + '://' + req.get('host');
-  }
 }
