+++ core-libs/setup/ssr/providers/index.ts
@@ -1,0 +1,1 @@
+export { getServerRequestProviders } from './ssr-providers';
