+++ core-libs/setup/ssr/providers/ssr-providers.ts
@@ -1,0 +1,25 @@
+import { StaticProvider } from '@angular/core';
+import { REQUEST } from '@nguniversal/express-engine/tokens';
+import { SERVER_REQUEST_ORIGIN, SERVER_REQUEST_URL } from '@spartacus/core';
+import { getRequestOrigin } from '../util/request-origin';
+import { getRequestUrl } from '../util/request-url';
+
+/**
+ * Returns Spartacus providers to be passed to the Angular express engine (in SSR)
+ *
+ * @param options
+ */
+export function getServerRequestProviders(): StaticProvider[] {
+  return [
+    {
+      provide: SERVER_REQUEST_URL,
+      useFactory: getRequestUrl,
+      deps: [REQUEST],
+    },
+    {
+      provide: SERVER_REQUEST_ORIGIN,
+      useFactory: getRequestOrigin,
+      deps: [REQUEST],
+    },
+  ];
+}
