+++ core-libs/setup/ssr/public_api.ts
@@ -1,2 +1,3 @@
 export * from './engine-decorator/index';
 export * from './optimized-engine/index';
+export * from './providers/index';
