+++ core-libs/setup/ssr/util/request-url.ts
@@ -1,0 +1,6 @@
+import { Request } from 'express';
+import { getRequestOrigin } from './request-origin';
+
+export function getRequestUrl(req: Request): string {
+  return getRequestOrigin(req) + req.originalUrl;
+}
